update for EncAPI and add API test sample code

+++ bjunion_enc/enc_api.h
@@ -5,14 +5,13 @@
 #include "enc_define.h"
 
 typedef void* (*AMVencHWFunc_Init)(int fd, amvenc_initpara_t* init_para);
-typedef AMVEnc_Status (*AMVencHWFunc_InitFrame)(void *dev, unsigned* yuv, AMVEncBufferType type, AMVEncFrameFmt fmt, bool IDRframe);
-typedef AMVEnc_Status (*AMVencHWFunc_EncodeSPS_PPS)(void *dev, unsigned char* outptr,int* datalen);
-typedef AMVEnc_Status (*AMVencHWFunc_EncodeSlice)(void *dev, unsigned char* outptr,int* datalen);
+typedef AMVEnc_Status (*AMVencHWFunc_InitFrame)(void *dev, ulong *yuv, AMVEncBufferType type, AMVEncFrameFmt fmt, bool IDRframe);
+typedef AMVEnc_Status (*AMVencHWFunc_EncodeSPS_PPS)(void *dev, unsigned char* outptr, int* datalen);
+typedef AMVEnc_Status (*AMVencHWFunc_EncodeSlice)(void *dev, unsigned char* outptr, int* datalen);
 typedef AMVEnc_Status (*AMVencHWFunc_CommitEncode)(void *dev, bool IDR);
 typedef void (*AMVencHWFunc_Release)(void *dev);
 
-typedef struct tagAMVencHWFuncPtr
-{
+typedef struct tagAMVencHWFuncPtr {
     AMVencHWFunc_Init Initialize;
     AMVencHWFunc_InitFrame InitFrame;
     AMVencHWFunc_EncodeSPS_PPS EncodeSPS_PPS;
@@ -21,15 +20,13 @@ typedef struct tagAMVencHWFuncPtr
     AMVencHWFunc_Release Release;
 } AMVencHWFuncPtr;
 
-
 typedef void* (*AMVencRCFunc_Init)(amvenc_initpara_t* init_para);
 typedef AMVEnc_Status (*AMVencRCFunc_PreControl)(void *dev, void *rc, int frameInc, bool force_IDR);
 typedef AMVEnc_Status (*AMVencRCFunc_PostControl)(void *dev, void *rc, bool IDR, int* skip_num, int numFrameBits);
-typedef AMVEnc_Status (*AMVencFunc_InitFrameQP)(void *dev, void *rc,bool IDR, int bitrate, float frame_rate);
+typedef AMVEnc_Status (*AMVencFunc_InitFrameQP)(void *dev, void *rc, bool IDR, int bitrate, float frame_rate);
 typedef void (*AMVencRCFunc_Release)(void *rc);
 
-typedef struct tagAMVencRCFuncPtr
-{
+typedef struct tagAMVencRCFuncPtr {
     AMVencRCFunc_Init Initialize;
     AMVencRCFunc_PreControl PreControl;
     AMVencRCFunc_PostControl PostControl;
@@ -44,9 +41,9 @@ extern AMVEnc_Status AMRCInitFrameQP(amvenc_hw_t* hw_info, bool IDR, int bitrate
 extern void AMCleanupRateControlModule(amvenc_hw_t* hw_info);
 
 extern AMVEnc_Status InitAMVEncode(amvenc_hw_t* hw_info, int force_mode);
-extern AMVEnc_Status AMVEncodeInitFrame(amvenc_hw_t* hw_info, unsigned* yuv, AMVEncBufferType type, AMVEncFrameFmt fmt, bool IDRframe);
-extern AMVEnc_Status AMVEncodeSPS_PPS(amvenc_hw_t* hw_info, unsigned char* outptr,int* datalen);
-extern AMVEnc_Status AMVEncodeSlice(amvenc_hw_t* hw_info, unsigned char* outptr,int* datalen);
+extern AMVEnc_Status AMVEncodeInitFrame(amvenc_hw_t* hw_info, ulong *yuv, AMVEncBufferType type, AMVEncFrameFmt fmt, bool IDRframe);
+extern AMVEnc_Status AMVEncodeSPS_PPS(amvenc_hw_t* hw_info, unsigned char* outptr, int* datalen);
+extern AMVEnc_Status AMVEncodeSlice(amvenc_hw_t* hw_info, unsigned char* outptr, int* datalen);
 extern AMVEnc_Status AMVEncodeCommit(amvenc_hw_t* hw_info, bool IDR);
 extern void UnInitAMVEncode(amvenc_hw_t* hw_info);
 #endif
